+++ lib/3scale/api/client.rb
@@ -28,8 +28,9 @@ module ThreeScale
       # @api public
       # @return [Array<Hash>]
       # @param [Fixnum] service_id Service ID
-      def list_applications(service_id)
-        response = http_client.get("/admin/api/applications", params: { service_id: service_id })
+      def list_applications(service_id: nil)
+        params = service_id ? { service_id: service_id } : nil
+        response = http_client.get("/admin/api/applications", params: params)
         extract(collection: 'applications', entity: 'application', from: response)
       end